Added first unit tests, exit on non 200
Now exits if request doesn't return 200. Fell over in a heap
much later on and confusingly without it. Added unit test to
cover this.

Unsure as to how to best check that BS object is returned
from the request function. Should I even be returning this
or should I be passing around the response object and then
parsing it in each downstream function?